Version 1.1.0 - Fixed severe initialization bug.
This library was implemented at first instance due to be used as a
console tool and worked perfetly until I tried to use as a library to
parse more than single file.

There was an insane (and obvious bug): I inadvertently left the expat
instantiation outside of the main (x2j parser) function so if you try to
use it just a second time, it were trying to feed it (including new xml
header) to the already opened expat parser instance so it were logically
throwing.

I decided to increase minor version number instead of patch because I'm
not adding new functonality, this version removes a really limiting bug
that was making impossible to use this library to parse more than one
document per program execution.